C++ compiler portability changes to files in utils folder (except for 1, 2, and 3 below). 1. Matrix::appendColumnToVector starts by resizing the     destination vector (so as to avoid reallocations). 2. hammingdistance.h uses intptr_t for sequence lengths     et cetera, so it should now work for sequences     of >2 billion nucleotides. 3. Params::gurobi_threads is now int rather than bool. 4. static_cast<int> used to silence conversion warnings     about implicit conversions. 5. In MS builds, Optimization::minimizeNewtonMulti     replaces dynamically-sized arrays on the stack with     boost::scoped_array instances (the former is     non-standard C++ and it prevented compilation     with the Microsoft compiler). 6. Calls to strcpy() are, where possible, replaced with calls     to strcpy_s().

+++ utils/checkpoint.cpp
@@ -337,7 +337,7 @@ void Checkpoint::endList() {
 }
 
 void Checkpoint::getSubCheckpoint(Checkpoint *target, string partial_key) {
-    int len = partial_key.length();
+    int len = static_cast<int>(partial_key.length());
     for (auto it = lower_bound(partial_key); it != end() && it->first.substr(0, len) == partial_key; it++) {
         target->put(it->first.substr(len+1), it->second);
     }
@@ -354,7 +354,7 @@ void Checkpoint::putSubCheckpoint(Checkpoint *source, string partial_key) {
 }
 
 void Checkpoint::transferSubCheckpoint(Checkpoint *target, string partial_key, bool overwrite) {
-    int len = partial_key.length();
+    int len = static_cast<int>(partial_key.length());
     for (auto it = lower_bound(partial_key); it != end() && it->first.substr(0, len) == partial_key; it++) {
         if (overwrite || !target->hasKey(it->first))
             target->put(it->first, it->second);
